Fix some compiler warnings
When compiling under Fedora 24 with Glibc 2.23.1 we were getting several
compiler warnings like

In file included from /usr/include/stdio.h:27:0,
                 from xlander.h:23,
                 from xlander.c:15:
/usr/include/features.h:148:3: warning: #warning "_BSD_SOURCE and
_SVID_SOURCE are deprecated, use _DEFAULT_SOURCE" [-Wcpp]
 # warning "_BSD_SOURCE and _SVID_SOURCE are deprecated, use _DEFAULT_SOURCE"
   ^~~~~~~

This is due to _BSD_SOURCE and _SVID_SOURCE being deprecated in Glibc
2.20

The simple fix is to just define _DEFAULT_SOURCE in xlander.h which is
included in all places where this warning was coming from. Using
_DEFAULT_SOURCE ensures that the interfaces provided by the above will
still be available.

Signed-off-by: Andrew Clayton <andrew@digital-domain.net>