Fix build issue with audio-null.c
On GitHub, @JoesCat reported an issue when building with audio-null.c
rather than audio-linux.c whereby it would fail to build with the
following linker error

  /usr/bin/ld: resource.o: in function `LoadResources':
  resource.c:(.text+0x2b9): undefined reference to `USE_AUDIO'
  collect2: error: ld returned 1 exit status

This was due to USE_AUDIO being declared as extern in globals.h but not
being defined in audio-null.c

We could of simply added a definition for it to audio-null.c, but it
isn't actually needed there, in fact currently the only place it _is_
needed is in audio-linux.c.

Given the above, lets confine USE_AUDIO only to audio-linux.c (if we get
any other audio-OS.c files, they can simply define it there also).

We then create a new snd_no_audio(void) function to disable the audio if
desired and call that in resources.c instead of messing with USE_AUDIO
directly.

Link: <https://github.com/ac000/xlander/pull/3>
Reported-by: Joe Da Silva <digital@joescat.com>
Signed-off-by: Andrew Clayton <andrew@digital-domain.net>